+++ services/Directory/FilterLists.Directory.Infrastructure/Persistence/Queries/Entities/Dependent.cs
@@ -7,9 +7,9 @@ namespace FilterLists.Directory.Infrastructure.Persistence.Queries.Entities;
 
 public record Dependent
 {
-    public int DependencyFilterListId { get; init; }
+    public long DependencyFilterListId { get; init; }
     public FilterList DependencyFilterList { get; init; } = null!;
-    public int DependentFilterListId { get; init; }
+    public long DependentFilterListId { get; init; }
     public FilterList DependentFilterList { get; init; } = null!;
 }
